+++ src/reader/reader.jl
@@ -33,6 +33,7 @@ function WeaveDoc(source, informat = nothing)
         path,
         chunks,
         "",
+        "",
         nothing,
         "",
         "",